Enhance statistics handling in cohort analysis
- Introduced a new `StatConfig` dataclass for configuration of statistics calculations.
- Added `format_stats_table` and `get_stratified_stats` functions to improve the formatting and stratification of statistics.
- Refactored `analyze_cohort` to utilize the new functions for better organization and clarity.
- Removed redundant code and streamlined the statistics calculation process.
